Move check*Event funs from Teams to Utils
